+++ src/fp/addBusinessDays/index.js
@@ -1,8 +1,8 @@
 // This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
 
-import fn from '../../addWeekDays/index.js'
+import fn from '../../addBusinessDays/index.js'
 import convertToFP from '../_lib/convertToFP/index.js'
 
-var addWeekDays = convertToFP(fn, 2)
+var addBusinessDays = convertToFP(fn, 2)
 
-export default addWeekDays
+export default addBusinessDays